+++ nacc_attribute_deriver/utils/date.py
@@ -4,11 +4,11 @@ import re
 from datetime import date, datetime
 from typing import List, Optional, Set, Tuple, Union
 
+from .errors import AttributeDeriverError
+
 
 def datetime_from_form_date(date_string: Optional[str]) -> Optional[datetime]:
     """Converts date string to datetime based on format.
-
-    Expects either `%Y-%m-%d` or `%m/%d/%Y`.
 
     Args:
       date_string: the date string
@@ -18,10 +18,27 @@ def datetime_from_form_date(date_string: Optional[str]) -> Optional[datetime]:
     if not date_string:
         return None
 
-    if re.match(r"\d{4}-\d{2}-\d{2}", date_string):
-        return datetime.strptime(date_string, "%Y-%m-%d")
+    try:
+        # YYYY-MM-DD format
+        if re.match(r"\d{4}-\d{2}-\d{2}", date_string):
+            return datetime.strptime(date_string, "%Y-%m-%d")
 
-    return datetime.strptime(date_string, "%m/%d/%Y")
+        # YYYY/MM/DD format
+        elif re.match(r"\d{4}/\d{2}/\d{2}", date_string):
+            return datetime.strptime(date_string, "%Y/%m/%d")
+
+        # MM-DD-YYYY
+        elif re.match(r"\d{2}-\d{2}-\d{4}", date_string):
+            return datetime.strptime(date_string, "%m-%d-%Y")
+
+        # MM/DD/YYYY
+        elif re.match(r"\d{2}/\d{2}/\d{4}", date_string):
+            return datetime.strptime(date_string, "%m/%d/%Y")
+
+        raise AttributeDeriverError(f"Invalid date format: {date_string}")
+
+    except ValueError as e:
+        raise AttributeDeriverError(f"Failed to parse date {date_string}: {e}") from e
 
 
 def date_from_form_date(date_string: Optional[str]) -> Optional[date]: