+++ CHANGELOG.md
@@ -9,9 +9,14 @@ Significant updates and refactorings to support:
 * rest of derived variables
 * missingness logic
 
-#### rc2
+#### RCs:
 
-* Adds V4 header variables
+* rc2: Adds V4 header variables
+* rc3:
+    * Updates B1a-related variables to not throw error when missing - data currently not available in FW
+    * Updates to handle NP form date being `npformdate` instead of `visitdate` in the new system
+    * Updates to not throw error if missing UDS in cross-module calculations, as it may have just been an MDS/BDS participant
+    * Updates to support `yyyy/mm/dd` and `mm-dd-yyyy` date formats
 
 ## 1.4.3